[Security] Test the concurrent usage of a signature hash in the component that owns SignatureHasher
