+++ src/Symfony/Component/Security/Core/composer.json
@@ -31,6 +31,7 @@
         "symfony/expression-language": "^4.4|^5.0|^6.0",
         "symfony/http-foundation": "^5.3|^6.0",
         "symfony/ldap": "^4.4|^5.0|^6.0",
+        "symfony/property-access": "^5.4|^6.0",
         "symfony/translation": "^5.4.35|~6.3.12|^6.4.3",
         "symfony/validator": "^5.2|^6.0",
         "psr/log": "^1|^2|^3"
